Move the contract subscription event handler under services
